+++ ChangeLog.md
@@ -8,6 +8,7 @@ Changes since last release
   - Add funtionality to add user-defined spotlights to the TiGLCreator ([#1245](https://github.com/DLR-SC/tigl/issues/1245))
 
 - Fixes
+  - Update modificator widget's references after undo/redo action. Fixes ([#1240](https://github.com/DLR-SC/tigl/issues/1240))
   - Fix wrong scaling of the main application, if the screen scale is set to 150% on Windows ([#1238](https://github.com/DLR-SC/tigl/issues/1238))
   - Add undo/redo buttons to toolbar ([#1123](https://github.com/DLR-SC/tigl/issues/1123))
   - Fix incorrect invalidation logic of elements, sections and fuselages/wings. This caused invalid geometries for some combinations of positionings and transformations when applying CPACS changes in TiGLCreator ([#1209](https://github.com/DLR-SC/tigl/issues/1209))